Hoist the style parts of parallel.rs into style/.
